git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@10258 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/Omicron/Omicron.cpp
@@ -960,7 +960,7 @@ bool Omicron::PharseAnswerString(std::string &InputBuffer, const std::string &Ko
 {
     std::string Seperator = "§";
 
-    unsigned int StrIndex = InputBuffer.find(Kommando);
+    size_t StrIndex = InputBuffer.find(Kommando);
     if( StrIndex != std::string::npos )
     {
         InputBuffer = InputBuffer.substr(StrIndex + Kommando.length());  // Erase command
